requirements.txt
requests>=2.25.1
python-chess>=1.9.0 